rename FC to FanController
