+++ Product/Product.vcxproj.filters
@@ -36,7 +36,7 @@
     <ClInclude Include="PressDetector.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="FC.h">
+    <ClInclude Include="FanController.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="Battery.h">
@@ -53,7 +53,7 @@
     <ClCompile Include="MySerial.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="FC.cpp">
+    <ClCompile Include="FanController.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="Battery.cpp">
